impl BitAnd, BitOr, BitXor UADF
